Switch to uint64
Goose only supports this integer type

+++ disk/disk.go
@@ -5,22 +5,22 @@ import "fmt"
 // Block is a 4096-byte buffer
 type Block []byte
 
-const BlockSize int = 4096
+const BlockSize uint64 = 4096
 
 // Disk provides access to a logical block-based disk
 type Disk interface {
 	// Read reads a disk block by address
 	//
 	// Expects a < Size().
-	Read(a int) Block
+	Read(a uint64) Block
 
 	// Write updates a disk block by address
 	//
 	// Expects a < Size().
-	Write(a int, v Block)
+	Write(a uint64, v Block)
 
 	// Size reports how big the disk is, in blocks
-	Size() int
+	Size() uint64
 
 	// Barrier ensures data is persisted.
 	//
@@ -33,7 +33,7 @@ type MemDisk []Block
 
 var _ Disk = MemDisk(nil)
 
-func NewMemDisk(numBlocks int) MemDisk {
+func NewMemDisk(numBlocks uint64) MemDisk {
 	disk := make([]Block, numBlocks)
 	for i := range disk {
 		disk[i] = make([]byte, BlockSize)
@@ -41,19 +41,19 @@ func NewMemDisk(numBlocks int) MemDisk {
 	return disk
 }
 
-func (d MemDisk) Read(a int) Block {
+func (d MemDisk) Read(a uint64) Block {
 	return d[a]
 }
 
-func (d MemDisk) Write(a int, v Block) {
-	if len(v) != BlockSize {
+func (d MemDisk) Write(a uint64, v Block) {
+	if uint64(len(v)) != BlockSize {
 		panic(fmt.Errorf("v is not block-sized (%d bytes)", len(v)))
 	}
 	d[a] = v
 }
 
-func (d MemDisk) Size() int {
-	return len(d)
+func (d MemDisk) Size() uint64 {
+	return uint64(len(d))
 }
 
 func (d MemDisk) Barrier() {}
@@ -66,17 +66,17 @@ func Init(d Disk) {
 }
 
 // Read (see the Disk documentation)
-func Read(a int) Block {
+func Read(a uint64) Block {
 	return implicitDisk.Read(a)
 }
 
 // Write (see the Disk documentation)
-func Write(a int, v Block) {
+func Write(a uint64, v Block) {
 	implicitDisk.Write(a, v)
 }
 
 // Size (see the Disk documentation)
-func Size() int {
+func Size() uint64 {
 	return implicitDisk.Size()
 }
